+++ gammapy/scripts/detect.py
@@ -1,6 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
+import logging
+log = logging.getLogger(__name__)
 from ..utils.scripts import get_parser
 
 __all__ = ['detect']
@@ -22,9 +24,8 @@ def detect(infile,
 
     TODO: explain.
     """
-    import logging
-    logging.basicConfig(level=logging.DEBUG, format='%(levelname)s - %(message)s')
     # from gammapy import detect
 
     raise NotImplementedError
     # TODO: implement me
+    log.info('Reading {}'.format(infile))