+++ gammapy/scripts/reflected_regions.py
@@ -1,7 +1,9 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
-from ..utils.scripts import get_parser
+import logging
+log = logging.getLogger(__name__)
+from ..utils.scripts import get_parser, set_up_logging_from_args
 
 __all__ = ['reflected_regions']
 
@@ -27,7 +29,11 @@ def main(args=None):
                         '[default=%(default)s]')
     parser.add_argument('--min_on_distance', type=float, default=0.1,
                         help='Minimum distance to the on region (deg)')
+    parser.add_argument("-l", "--loglevel", default='info',
+                        choices=['debug', 'info', 'warning', 'error', 'critical'],
+                        help="Set the logging level")
     args = parser.parse_args(args)
+    set_up_logging_from_args(args)
     reflected_regions(**vars(args))
 
 
@@ -44,18 +50,23 @@ def reflected_regions(x_on,
 
     TODO: explain a bit.
     """
-    import logging
-    logging.basicConfig(level=logging.DEBUG, format='%(levelname)s - %(message)s')
     from astropy.io import fits
     from gammapy.background import ReflectedRegionMaker
 
-    logging.info('Reading {0}'.format(exclusion))
-    exclusion = fits.open(exclusion)[0]
+    if exclusion:
+        log.info('Reading {0}'.format(exclusion))
+        exclusion = fits.open(exclusion)[0]
+    else:
+        # log.info('No exclusion mask used.')
+        # TODO: make this work without exclusion mask
+        log.error("Currently an exclusion mask is required")
+        exit(-1)
 
     fov = dict(x=x_fov, y=y_fov, r=r_fov)
     rr_maker = ReflectedRegionMaker(exclusion=exclusion,
                                     fov=fov)
     source = dict(x_on=x_on, y_on=y_on, r_on=r_on)
     rr_maker.compute(**source)
-    logging.info('Writing {0}'.format(outfile))
+
+    log.info('Writing {0}'.format(outfile))
     rr_maker.write_off_regions(outfile)